# docs/autogenerated/openapi.yaml
openapi: 3.0.1
info:
  title: REST API
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
- url: https://localhost:8080
  description: Generated server url
...
paths:
  /repair-management/v2/repairs:
    get:
      tags:
      - Repair-Management
      summary: Get manual repairs.
      description: Get manual repairs which are running/completed/failed.
      operationId: get-repairs
      parameters:
      - name: keyspace
        in: query
        description: "Only return repairs matching the keyspace, mandatory if 'table'\
          \ is provided."
        required: false
        schema:
          type: string
      - name: table
        in: query
        description: Only return repairs matching the table.
        required: false
        schema:
          type: string
      - name: hostId
        in: query
        description: Only return repairs matching the hostId.
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OnDemandRepair'
    post:
      tags:
      - Repair-Management
      summary: Run a manual repair.
      description: "Run a manual repair, if 'isLocal' is not provided this will trigger\
        \ a cluster-wide repair."
      operationId: trigger-repair
      parameters:
      - name: keyspace
        in: query
        description: "The keyspace to run repair for, mandatory if 'table' is provided."
        required: false
        schema:
          type: string
      - name: table
        in: query
        description: The table to run repair for.
        required: false
        schema:
          type: string
      - name: isLocal
        in: query
        description: "Decides if the repair should be only for the local node, i.e\
          \ not cluster-wide."
        required: false
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OnDemandRepair'
...
  /repair-management/v2/repairs/{id}:
    get:
      tags:
      - Repair-Management
      summary: Get manual repairs matching the id.
      description: Get manual repairs matching the id which are running/completed/failed.
      operationId: get-repairs-by-id
      parameters:
      - name: id
        in: path
        description: Only return repairs matching the id.
        required: true
        schema:
          type: string
      - name: hostId
        in: query
        description: Only return repairs matching the hostId.
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OnDemandRepair'
...
components:
  schemas:
    OnDemandRepair:
      required:
      - completedAt
      - hostId
      - id
      - keyspace
      - repairedRatio
      - status
      - table
      type: object
      properties:
        id:
          type: string
          format: uuid
        hostId:
          type: string
          format: uuid
        keyspace:
          type: string
        table:
          type: string
        status:
          type: string
          enum:
          - COMPLETED
          - IN_QUEUE
          - WARNING
          - ERROR
          - BLOCKED
        repairedRatio:
          maximum: 1
          minimum: 0
          type: number
          format: double
        completedAt:
          minimum: -1
          type: integer
          format: int64
...